api/requirements.txt

fastapi>=0.103.1
uvicorn>=0.23.2
stripe>=6.5.0
supabase>=1.0.3
python-dotenv>=1.0.0
